big-bang: Lean-gate polarity INVERSION (Lean unconditional; one no-lean-link platform gate) + triage bulk test-greening
INVERSION (build-infra, approved): the seven per-crate Lean feature gates
(captp/coord lean-gate, federation lean-admission, turn lean-shadow,
intent verified-settle, sdk/app-framework lean-producer) are DELETED.
Lean verdicts/producers/differentials are now the unconditional default
on native; the single workspace feature no-lean-link (OFF by default)
exists only for targets that cannot link libdregg_lean.a (wasm32/zkvm,
plus a build.rs target hard-skip as defense-in-depth). node/Cargo.toml
drops all five feature enables — absence of verification is now visible
in the build graph. Verified: workspace native cargo check green (9
crates, exit 0) AND dregg-wasm wasm32-unknown-unknown green (exit 0).

TRIAGE GREENING (Opus lane, mirrored green on persvati, 6888 tests):
signed-wells i64 straggler sweep across test files; dead tests deleted
(captp_effects_pipeline, cross_federation_captp_turn, proof_obligation
example); coverage_state_constraints added; ratchet to 9.

+++ coord/src/entangled_diff.rs
@@ -137,10 +137,12 @@ mod atomicity_diff {
         (seed, pk)
     }
 
+    // signed-wells (ac01f9b7b): cell balances are i64; keep a u64 convenience
+    // param (callers pass non-negative bal0/bal1/bal2) and convert at the boundary.
     fn permissive_cell(key_byte: u8, balance: u64) -> Cell {
         let mut pk = [0u8; 32];
         pk[0] = key_byte;
-        let mut cell = Cell::with_balance(pk, [0u8; 32], balance);
+        let mut cell = Cell::with_balance(pk, [0u8; 32], balance as i64);
         cell.permissions = dregg_cell::Permissions {
             send: dregg_cell::AuthRequired::None,
             receive: dregg_cell::AuthRequired::None,
